Moved containers and applications to handlers package
