+++ lighthouse.go
@@ -24,8 +24,8 @@ import (
     "github.com/lighthouse/lighthouse/beacons/aliases"
     "github.com/lighthouse/lighthouse/users"
     "github.com/lighthouse/lighthouse/handlers/docker"
-    "github.com/lighthouse/lighthouse/containers"
-    "github.com/lighthouse/lighthouse/containers/applications"
+    "github.com/lighthouse/lighthouse/handlers/containers"
+    "github.com/lighthouse/lighthouse/handlers/containers/applications"
 
     "github.com/lighthouse/lighthouse/logging"
